requirements.txt
Kivy>=2.2.0
django-environ>=0.10.0
kivymd>=1.1.1
